Refactor user queries
