+++ server/db/models/utilities/controllers.js
@@ -1,0 +1,7 @@
+module.exports = function(sequelize, User) {
+  var UserController = require('./userController')(sequelize, User);
+
+  return {
+    UserController: UserController
+  }
+}
